requirements.txt
mypy
wheel
twisted
pytz
pyyaml
sqlalchemy
mysqlclient
pyopenssl
service_identity
PyCryptodome
inflection
coloredlogs
pylibmc